refactor: 스트리밍 서비스 데이터베이스 타입을 DatbaseType 파일에 옮겼다.
